Switch from .rvmrc to ruby-version
.rvmrc is an old way of annotating ruby projects. The file is
a shell script that requires trusting, and, therefore
triggers a warning every time it is changed, interactively
asking whether the user trusts it. This is not ideal inside a
Docker container, where all processes should be
non-interactive.

The solution, as suggested by rvm, is to switch to a
ruby-version file. This commit makes the switch and ensures
that selenium-cucumber still works.