Rewrite run.sh file and rename it to qata
For remote execution of tests inside a VM, the Docker container is
now run daemonized with a service interface. Tests can be executed
by requesting a link on the listening port. To reflect this change
in execution model, the run.sh command has acquired a service
interface (start, stop, restart).

A run file is kept at /var/run to store the ID of the Docker
container when it is started. This is used to stop and remove the
container.

Additionally, run.sh has been renamed to qata to reflect the
change in its function.