Decide on naming conventions for tests execution
Since, tests execution results will pass through two servers
before being served on the dashboard, it is important to have
a set naming convention for different objects, to be used across
both servers and frontend.

An execution of a set of tests can be requested. The result of
this execution are stored in TestsExecResult which is managed by
an instance of TestsExecResultsManager.